refactor(下载任务页面):重构下载任务展示方式
-将 DataGrid 替换为 ListBox，使用自定义模板展示任务信息
- 添加已下载片段数量属性和进度条显示
- 修改任务进度计算方式，使用 double 类型
- 优化任务列表样式，增加鼠标双击事件

+++ m3u8_downloader/Models/TaskProgress.cs
@@ -5,6 +5,6 @@
         public int TotalSegments { get; set; }
         public int DownloadedSegments { get; set; }
         public long TotalBytes { get; set; }
-        public int PercentComplete { get; set; }
+        public double PercentComplete { get; set; }
     }
 }
